+++ release-notes.md
@@ -1,5 +1,6 @@
 # Release 1.0.6
 
+- Fix type conversion in warpBitmapWithCorners
 - Ignore release APK
 - Add OpenCV-based document scanner with CameraX
 - Add OpenCV-based document scanner with CameraX